Add queries route and DBHandler
(queries.js) : See app.js to see that queries.js is mounted to the path '/query'; So, if the user types in {servername}/query, then express will use what is found in queries.js
queries.js uses DBHandler to send all of the events in the Events table back to the front-end.
(DBHandler.js) : This file contains some base functions for viewing the database. Another file just needs to 'require(DBHandler)' in order to use its functions. See queries.js