fix(Marker): Move display calculation to constructor
I noticed quite a bit of lag and the occasional React error when zooming in and out a lot. Turns out the animation delay was being recalculated many times every time the map moved. Obviously this is unnecessary, so the display data is now calculated just once, when the Marker loads.
For an updated display, the user will need to reload the page (a price I'm willing to pay).